Avoid infinite loop when reading assembly file (#18)
When reading in shared data from an assembly file, the loop (weirdly...) uses a
left shift instead of a normal iteration when determining the assembly
datatype.  However, it started the loop variable at 0, so it would remain at 0
after a left shift and loop infinitely.  This sets the loop variable to 1,
which seems to have been the intention since the cGasTypes contains empty dummy
entries to accommodate the shifting method.